remove promotional qr screen, all in the same screen
